# openapi.yaml
openapi: 3.1.0
info:
  title: Groceries API
  version: 1.0.0
  description: |
    REST API for the Groceries application. Provides token-based authentication
    suitable for native app clients and future frontend rewrites.

    ## Authentication

    All endpoints except `POST /api/v1/auth/login` require a Bearer token in the
    `Authorization` header:

    ```
    Authorization: Bearer <token>
    ```

    Tokens are obtained from the login endpoint and expire after 720 hours (30 days).
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
          examples:
            - "item not found"

    # --- Auth ----------------------------------------------------------------
...
    User:
      type: object
      required: [id, name, admin]
      properties:
        id:
          type: integer
          examples:
            - 1
        name:
          type: string
          examples:
            - "alice"
        admin:
          type: boolean
          description: Whether this user has administrator privileges
...
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          examples:
            - "alice"
        admin:
          type: boolean

    # --- Group ---------------------------------------------------------------
...
    Store:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          examples:
            - 1
        name:
          type: string
          minLength: 3
          examples:
            - "Whole Foods"
...
    UpdateStoreRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 3
          examples:
            - "Whole Foods"

    # --- Category ------------------------------------------------------------
    Category:
      type: object
      required: [id, store_id, name, description, item_count]
      properties:
        id:
          type: integer
          examples:
            - 1
        store_id:
          type: integer
          examples:
            - 1
        name:
          type: string
          minLength: 3
          examples:
            - "Produce"
        description:
          type: string
          examples:
            - "Fresh fruits and vegetables"
        item_count:
          type: integer
          description: Number of items assigned to this category
          examples:
            - 12
...
  responses:
    NoContent:
      description: Operation succeeded with no response body
    BadRequest:
      description: The request was malformed or failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The authenticated user does not have permission for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The operation conflicts with existing data (e.g. duplicate name)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: An unexpected server-side error occurred
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Numeric resource ID
...
security:
  - bearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/users/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"

    get:
      operationId: getUser
      summary: Get a user by ID
      tags: [users]
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      operationId: updateUser
      summary: Update a user
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      operationId: deleteUser
      summary: Delete a user
      tags: [users]
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # --------------------------------------------------------------------------
  # Groups
  # --------------------------------------------------------------------------
...
  /api/v1/stores/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"

    get:
      operationId: getStore
      summary: Get a store by ID, including its categories
      tags: [stores]
      responses:
        "200":
          description: Store with categories
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Store"
                  - type: object
                    required: [categories]
                    properties:
                      categories:
                        type: array
                        items:
                          $ref: "#/components/schemas/Category"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      operationId: updateStore
      summary: Update a store
      tags: [stores]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoreRequest"
      responses:
        "200":
          description: Updated store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      operationId: deleteStore
      summary: Delete a store
      description: Fails if any categories are still assigned to the store.
      tags: [stores]
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # --------------------------------------------------------------------------
  # Categories
  # --------------------------------------------------------------------------
...